src/flask_src/requirements.txt
blinker==1.9.0
click==8.1.7
Flask==3.1.0
flask-marshmallow==1.2.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
marshmallow==3.23.1
marshmallow-sqlalchemy==1.1.0
mysql-connector-python==9.1.0
packaging==24.2
python-dotenv==1.0.1
SQLAlchemy==2.0.36
typing_extensions==4.12.2
Werkzeug==3.1.3
